move hook to app.py
