+++ passenger_wsgi.py
@@ -1,5 +1,6 @@
-import os
-import sys
+# coding=utf8
+# Copyright (c) 2019 CineUse
+
 import app
 
 application = app.app
